#314 comment: I believe I found the big, bad bug. I did my testing on the cone and I changed it to NOT use the osg geometry that was being shown to create the collision geometry. It creates an entirely seperate osg geometry and uses that. This prevents it from being associated with the other translation matrices in the scene. I beleive the vortex stuff was somehow using that when creating the verices, causing the psoition to be incorrect.
I still have to go back and fix MeshBase though. However, to do this it just made sense to fix the other bug where I was not showing the actual collision mesh, but was instead showing the mesh that was loaded in. I am going to have to change the meshbase code so that if it is a collision object it creates the graphics by first creating the collision geometry and then creating an osg display node using that. If it is graphics mesh then it loads it in like normal. I went ahead and wrote a simple method that takes the ConvexMesh and creates a  basic osg node using its polygons. This shows the mesh from the collision object. I will use this one for the graphics of a collision object. For trianglemesh I will just use the loaded in version and disable the texture or something. This is fine since the collision mesh should match the loaded mesh one for one.