I converted the project to use the latest 5.1 version of vortex. I also added code to show the mass and volume of rigid body parts. I still need to write a test for that. I also fixed a few bugs where it was not passing back a pointer correctly.
I added the abililty to set a scale value for the simulation stability params. This scale value is set on a new project based on the unit scaling values.

Forced planes to be frozen.

Fixed some problems with the testing. I added code to allow the testing system to check if the data charts had been written out before it attempted to do a comparison.

Fixed the problem with the spring/enable disable problem