Fixed a bug in the scripted sim window where it was saving out the default camera position as position. The base sim window was loading this in as the position of the screen. When the camera position was negative this caused an error. I changed it to default position instead.
Fixed another bug where the base description of animatform was nothing. I defaulted it to blank and set the description for the scripted sim window. This was causing an error when showing that form in the tool selector dialog.

I also increased the timeout for doing a simulation block. It was causing an error while trying to open up windows on big simulations.