Added constraint limit damping/stiffness. This appears to work pretty well for the prismatic joint, but does not appear to work at all for hinge joints. I modified the GUI hinge stuff to no longer show separate upper/lower constraint objects for a joint. I did this because bullet does not have the ability to set stiffness/damping params for individual limits, so it makes no sense. I also removed the constraint relaxation stuff from being visible in bullet since I could not find any way to actually control this in Bullet at the moment.
I also started working on constraint friction. I believe I have a basic version working on the hinge joint that qualitatively reproduces the behavior in vortex. I still need to get it to update the friction values when changing in the GUI though. Right now you have to enable/disable the motor for it to work.