I got the basics of the constraint friction working for the hinge. I also fixed an issue with the reporting of joint position and velocity of hinge joints. I modified it so it keeps track of the joint position internally and just uses a delta of the prev and current position reported by bullet. This allows me to wrap the joint position passed 2PI so position and velocity have no discontinuites.
I did run into one serious issue with joint friction. The method I am using for it did not work at all with the slider joint. Setting the motor on, even with a very low maximum force, prevented the part from moving at all when applying a force to the child part. I believe I am going to have to go a different route to get this to work.