I was getting some compile problems with the new stlsoft performance_counter code on ubuntu. It was not able to use int64 and was trying to cast it to timeval, which was not compiling. So I switched this to just use the stlsoft define of the epochtype instead since they already dealt with the cross-platform compatability issues.
I also made a modification to not use boost.thread.try_join_for method if we are using boost version less than 1.50. That method was not added until that version. Instead it will just use a regular join and not worry about time outs.

I also recreated all the project files for ubuntu.