RoboticsAnimatSim runs on Odroid
I was able to get RoboticsAnimatSim to run on the odroid. I replaced the
platformstl code and switched back to using osg and openthreads. The
only reason I was using platformstl before was because I had trouble
getting those two to work on odroid. However, since I was having trouble
getting platformstl to work for osg on ubuntu I decided to try getting
osg::timer and openthreads to work on odroid again and fixed it.